requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
python-dotenv
openai
tiktoken
langchain-text-splitters
google-generativeai
matplotlib
transformers
accelerate
torch
ollama
huggingface_hub
hf_xet